commit4

+++ shop/src/main/java/com/example/shop/entity/dto/CategoryDto.java
@@ -1,8 +1,10 @@
 package com.example.shop.entity.dto;
 
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 
 @Data
 @NoArgsConstructor
@@ -11,6 +13,6 @@ public class CategoryDto {
 
     private Integer categoryId;
     private String categoryTitle;
-    private Integer numberOfProduct;
+    private Long numberOfProduct;
 
 }
